requirements.txt
crewai==0.12.1
beautifulsoup4==4.12.3
requests==2.31.0
pandas==2.2.0
numpy==1.26.3
selenium==4.16.0
webdriver_manager==4.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
openai==1.10.0
langchain==0.1.0